+++ src/lib/alienvault-pulse.js
@@ -7,7 +7,7 @@ const axios = require("axios");  // Importing axios for HTTP requests
 const fs = require("fs");  // Importing fs for file system operations
 const path = require("path");  // Importing path for file path manipulation
 const { config }  = require('../lib/configLoader');  // Importing the config for API keys
-const cron = require("node-cron");  // Importing cron for scheduling tasks
+const cron = require("node-cron");  // node-cron v4
 
 const API_KEY = config.alienvault.key;  // Accessing the API key from the config
 const OTX_URL = "https://otx.alienvault.com/api/v1/pulses/subscribed";  // URL for fetching pulses